Use twisted.internet.threads.deferToThread in the obfs3 handshake.
This cuts the amount of time that is spent doing UniformDH related work in the
context of the main event loop in half.  It is impossible to eliminate it
entirely as the client and server both send the public keys immediately on
connection when doing an obfs3 handshake.

Notable changes:
 * self.dh.get_secret() is processed in a Twisted worker thread.
 * Another state is added as part of the handshake process for when the thread
   is running.
 * At each step of the handshake process, if there is data remaining unprocessed
   the next stage is direcly invoked at the tail of each routine.
   * After the handshake is done, if data is pending, self._scan_for_magic()
     is called.
   * After the magic is found, if data is pending, we relay the remaining data
     immediately.