Clean up error handling when the http proxy doesn't return status.
Some HTTP proxys opt to close the connection instead of returning a
HTTP status code on certain failures (Eg: polipo auth faliure, privoxy
ACL denial).  This change logs a better error message to the obfsproxy
log when that happens.

As an added bonus the SOCKS5 server now knows about EHOSTUNREACH, so
a more informative response is sent on that failure type.